Declare action in form field instead of URL path
The benefit using an action form field instead of
defining the action in the path is that the URL in
the browser does not change. This is particularly
noticable when a form POST returns an error where
the client (i.e. broswer) is not redirected using
the post/redirect/get approach.

https://en.wikipedia.org/wiki/Post/Redirect/Get